# routes/installations.js
const express = require('express');
const router = express.Router();
const crypto = require('crypto');
const { SolarInstallation, GenerationReading, GridSubstation } = require('../models');
const { errorBody, sendError } = require('../utils/errors');
const userAuth = require('../middleware/userAuth');
const { installationScopeFilter } = require('../middleware/scopeByJurisdiction');
const deviceAuth = require('../middleware/deviceAuth');
const requireNational = require('../middleware/requireNational');
const { signDeviceToken, DEVICE_SCOPE } = require('../utils/tokens');
const {
  validateIdParam, validateCreateInstallation, validateReplaceInstallation,
} = require('../validation/schemas');

const sha256 = (text) => crypto.createHash('sha256').update(text).digest('hex');

// GET /installations — scoped by the logged-in user's jurisdiction
router.get('/', userAuth, async (req, res) => {
  try {
    const filter = await installationScopeFilter(req.user);

    const installations = await SolarInstallation.find(filter)
      .select('-apiKeyHash')
      .populate('substation', 'name code')
      .sort({ meterId: 1 });

    res.json(installations);
  } catch (err) {
    sendError(res, err, 'FETCH_FAILED', 'Could not fetch installations');
  }
});

// GET /installations/:id/last-reading — operational view: just the newest reading
router.get('/:id/last-reading', userAuth, validateIdParam, async (req, res) => {
  try {
    const installation = await SolarInstallation.findById(req.params.id);
    if (!installation) {
      return res.status(404).json(errorBody('NOT_FOUND', 'Installation not found'));
    }

    // jurisdiction gate: same rule as the collection endpoint
    const filter = await installationScopeFilter(req.user);
    const inScope = await SolarInstallation.findOne({ _id: req.params.id, ...filter }).select('_id');
    if (!inScope) {
      return res.status(403).json(errorBody('FORBIDDEN', 'This installation is outside your jurisdiction'));
    }

    const reading = await GenerationReading.findOne({ installation: req.params.id })
      .sort({ timestamp: -1 });

    if (!reading) {
      return res.status(404).json(errorBody('NO_READINGS', 'No readings yet for this installation'));
    }

    res.json(reading);
  } catch (err) {
    sendError(res, err, 'FETCH_FAILED', 'Could not fetch last reading');
  }
});


// GET /installations/:id — composite: the installation + its latest reading
router.get('/:id', userAuth, validateIdParam, async (req, res) => {
  try {
    const installation = await SolarInstallation.findById(req.params.id)
      .select('-apiKeyHash')
      .populate({
        path: 'substation',
        select: 'name code district',
        populate: { path: 'district', select: 'name code province', populate: { path: 'province', select: 'name code' } },
      });

    if (!installation) {
      return res.status(404).json(errorBody('NOT_FOUND', 'Installation not found'));
    }

    // jurisdiction gate
    const filter = await installationScopeFilter(req.user);
    const inScope = await SolarInstallation.findOne({ _id: req.params.id, ...filter }).select('_id');
    if (!inScope) {
      return res.status(403).json(errorBody('FORBIDDEN', 'This installation is outside your jurisdiction'));
    }

    const latestReading = await GenerationReading.findOne({ installation: installation._id })
      .sort({ timestamp: -1 });

    const payload = {
      ...installation.toObject(),
      latestReading: latestReading || null,
    };

    const body = JSON.stringify(payload);
    const etag = require('etag')(body);

    res.set('ETag', etag);
    res.set('Last-Modified', installation.updatedAt.toUTCString());

    const ifNoneMatch = req.headers['if-none-match'];
    if (ifNoneMatch && ifNoneMatch === etag) {
      return res.status(304).end();
    } 

    res.type('application/json').send(body);
  } catch (err) {
    sendError(res, err, 'FETCH_FAILED', 'Could not fetch installation');
  }
});

// POST /installations/:id/token — bootstrap: a device presents its long-lived
// API key (x-api-key) and receives a short-lived, installation-scoped JWT.
router.post('/:id/token', validateIdParam, deviceAuth, async (req, res) => {
  try {
    // short-lived access token; the API key stays the long-lived secret
    const token = signDeviceToken(req.params.id);
    res.json({ token, tokenType: 'Bearer', expiresIn: 3600, scope: DEVICE_SCOPE });
  } catch (err) {
    sendError(res, err, 'TOKEN_FAILED', 'Could not issue token');
  }
});

// POST /installations — provision a new solar installation (national-only).
router.post('/', userAuth, requireNational, validateCreateInstallation, async (req, res) => {
  try {
    const { meterId, capacityKw, latitude, longitude, substation } = req.valid.body;

    // the substation must exist, so we never orphan an installation
    const sub = await GridSubstation.findById(substation);
    if (!sub) {
      return res.status(400).json(errorBody('INVALID_SUBSTATION', 'substation does not exist'));
    }

    const apiKey = crypto.randomBytes(24).toString('hex'); // device secret, shown once
    const installation = await SolarInstallation.create({
      meterId, capacityKw, latitude, longitude, substation,
      apiKeyHash: sha256(apiKey), // we store only the hash
    });

    const safe = installation.toObject();
    delete safe.apiKeyHash; // never expose the stored hash

    res
      .status(201)
      .location(`${req.baseUrl}/${installation._id}`)
      .json({ ...safe, apiKey }); // apiKey returned once so the device can authenticate
  } catch (err) {
    if (err.code === 11000) { // duplicate meterId (unique index)
      return res.status(409).json(errorBody('DUPLICATE_METER', 'An installation with this meterId already exists'));
    }
    sendError(res, err, 'CREATE_FAILED', 'Could not create installation');
  }
});

// PUT /installations/:id — fully replace an installation's editable fields (national-only, idempotent)
router.put('/:id', userAuth, requireNational, validateReplaceInstallation, async (req, res) => {
  try {
    const { capacityKw, latitude, longitude, substation } = req.valid.body;

    const existing = await SolarInstallation.findById(req.params.id);
    if (!existing) {
      return res.status(404).json(errorBody('NOT_FOUND', 'Installation not found'));
    }

    // same rule as provisioning: never point an installation at a substation that doesn't exist
    if (!(await GridSubstation.exists({ _id: substation }))) {
      return res.status(400).json(errorBody('INVALID_SUBSTATION', 'substation does not exist'));
    }

    // optimistic concurrency: only proceed when the resource has NOT changed since the client fetched it
    const ifUnmodifiedSince = req.header('If-Unmodified-Since');
    if (ifUnmodifiedSince) {
      const since = new Date(ifUnmodifiedSince);
      const lastModifiedSec = Math.floor(existing.updatedAt.getTime() / 1000); // Last-Modified is second-precision
      const sinceSec = Math.floor(since.getTime() / 1000);
      if (isNaN(since.getTime()) || lastModifiedSec > sinceSec) {
        return res.status(412).json(
          errorBody('PRECONDITION_FAILED', 'Installation has changed since you last fetched it; re-fetch and retry')
        );
      }
    }

    const updated = await SolarInstallation.findByIdAndUpdate(
      req.params.id,
      { capacityKw, latitude, longitude, substation }, // meterId and apiKeyHash are NOT editable here
      { new: true, runValidators: true }
    ).select('-apiKeyHash');

    res.set('Last-Modified', updated.updatedAt.toUTCString());
    res.json(updated);
  } catch (err) {
    sendError(res, err, 'UPDATE_FAILED', 'Could not update installation');
  }
});

// DELETE /installations/:id — remove an installation (national-only)
router.delete('/:id', userAuth, requireNational, validateIdParam, async (req, res) => {
  try {
    const deleted = await SolarInstallation.findByIdAndDelete(req.params.id);
    if (!deleted) {
      return res.status(404).json(errorBody('NOT_FOUND', 'Installation not found'));
    }
    res.status(204).send(); // 204 = success, nothing to return
  } catch (err) {
    sendError(res, err, 'DELETE_FAILED', 'Could not delete installation');
  }
});

module.exports = router;