# swagger.yaml
openapi: 3.0.3
info:
  title: SLSEA Solar Generation Data API
  description: >
    Real-time and historical solar generation data API for the Sri Lanka
    Sustainable Energy Authority. Metering devices push readings as write
    clients; SLSEA users read data scoped by jurisdiction.
  version: 1.0.0
servers:
  - url: https://webapicw-production.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local development
...
paths:
  /provinces:
    get:
      tags: [Hierarchy]
      summary: List all provinces
      responses:
        '200':
          description: Array of provinces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Province'
        '406':
          $ref: '#/components/responses/NotAcceptable'
  /provinces/{id}:
    get:
      tags: [Hierarchy]
      summary: Get a single province (atomic resource)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: One province
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Province'
        '404':
          $ref: '#/components/responses/NotFound'
  /provinces/{id}/districts:
    get:
      tags: [Hierarchy]
      summary: List districts in a province (scoped collection)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Array of districts in this province
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/District'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /substations/{id}/installations:
    get:
      tags: [Substations]
      summary: List installations connected to this substation
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Array of installations at this substation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstallationBasic'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /installations:
    get:
      tags: [Installations]
      summary: List installations (scoped to the logged-in user's jurisdiction)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of installations visible to this user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Installation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
    post:
      tags: [Installations]
      summary: Provision a new installation (returns a one-time device API key)
      description: >
        Creates an installation and generates its device API key. The plain key
        is returned ONCE; only its hash is stored. The device uses that key at
        POST /installations/{id}/token to obtain write tokens.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [meterId, capacityKw, latitude, longitude, substation]
              properties:
                meterId: { type: string, minLength: 1, maxLength: 50 }
                capacityKw: { type: number, exclusiveMinimum: true, minimum: 0 }
                latitude: { type: number, minimum: -90, maximum: 90 }
                longitude: { type: number, minimum: -180, maximum: 180 }
                substation: { type: string, description: "Substation ObjectId" }
      responses:
        '201':
          description: Installation created (apiKey shown once)
          headers:
            Location:
              description: URL of the newly created installation
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Installation'
                  - type: object
                    properties:
                      apiKey:
                        type: string
                        description: Plain device API key — shown only once
        '400':
          description: Validation failed (VALIDATION_ERROR, field list in error.detail) or unknown substation (INVALID_SUBSTATION)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: An installation with this meterId already exists
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: National-role access required
  /installations/{id}:
    get:
      tags: [Installations]
      summary: Get one installation with its latest reading (composite resource)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Installation with latestReading embedded
          headers:
            ETag:
              description: Entity tag of this representation (use for conditional GET)
              schema: { type: string }
            Last-Modified:
              description: Last modification time (use as If-Unmodified-Since on PUT)
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationDetail'
        '304':
          description: Not Modified (If-None-Match matched the current ETag)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Installation is outside your jurisdiction
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Installations]
      summary: Replace an installation's editable fields (idempotent)
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: If-Unmodified-Since
          in: header
          required: false
          description: Optimistic concurrency — pass the Last-Modified from your last GET; the update is rejected with 412 if the resource changed since.
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capacityKw, latitude, longitude, substation]
              properties:
                capacityKw: { type: number, exclusiveMinimum: true, minimum: 0 }
                latitude: { type: number, minimum: -90, maximum: 90 }
                longitude: { type: number, minimum: -180, maximum: 180 }
                substation: { type: string }
      responses:
        '200':
          description: Updated installation (Last-Modified header returned)
        '400':
          description: Validation failed (VALIDATION_ERROR), malformed id (INVALID_ID) or unknown substation (INVALID_SUBSTATION)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          description: Precondition failed — the installation changed since If-Unmodified-Since
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: National-role access required
    delete:
      tags: [Installations]
      summary: Delete an installation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted successfully, no content
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: National-role access required
  /installations/{id}/token:
    post:
      tags: [Installations]
      summary: Exchange a device API key for a short-lived installation-write JWT
      description: >
        Bootstrap endpoint. The device authenticates with its long-lived API key
        and receives a short-lived JWT scoped to installation-write for this
        installation. That token is then used on POST /installations/{id}/readings.
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: A scoped bearer token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  tokenType: { type: string, example: Bearer }
                  expiresIn: { type: integer, example: 3600 }
                  scope: { type: string, example: installation-write }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: API key does not belong to this installation
        '404':
          $ref: '#/components/responses/NotFound'
  /installations/{id}/last-reading:
    get:
      tags: [Readings]
      summary: Most recent reading for an installation (derived/operational resource)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: The single most recent reading
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reading'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Installation is outside your jurisdiction
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        SLSEA USER token from POST /auth/login (valid 8 hours). Used for every
        read/admin endpoint. A device token is rejected here with
        403 WRONG_TOKEN_TYPE, even though it is validly signed.
    deviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        DEVICE token from POST /installations/{id}/token (valid 1 hour, scope
        installation-write, bound to one installation). Used only to push
        readings. A user token is rejected here with 403 WRONG_TOKEN_TYPE.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Device API key, issued at installation provisioning
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
      description: MongoDB ObjectId (24 hex characters). A malformed value is a 400 INVALID_ID.
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotAcceptable:
      description: The requested representation is not available; this API only produces application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Province:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
    District:
      type: object
      description: Shape returned by GET /provinces/{id}/districts — province is just an ObjectId reference
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        province: { type: string, description: "ObjectId reference to the parent province" }
...
    Installation:
      type: object
      description: Shape returned by GET /installations (list view)
      properties:
        _id: { type: string }
        meterId: { type: string }
        capacityKw: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        substation:
          type: object
          properties:
            _id: { type: string }
            name: { type: string }
            code: { type: string }
    InstallationDetail:
      type: object
      description: Shape returned by GET /installations/{id} — includes full substation hierarchy and latest reading
      properties:
        _id: { type: string }
        meterId: { type: string }
        capacityKw: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        substation:
          type: object
          properties:
            _id: { type: string }
            name: { type: string }
            code: { type: string }
            district:
              type: object
              properties:
                _id: { type: string }
                name: { type: string }
                code: { type: string }
                province:
                  type: object
                  properties:
                    _id: { type: string }
                    name: { type: string }
                    code: { type: string }
        latestReading:
          oneOf:
            - $ref: '#/components/schemas/Reading'
            - type: 'null'
    InstallationBasic:
      type: object
      description: Shape returned by GET /substations/{id}/installations — substation is just an ObjectId reference (already known from the URL)
      properties:
        _id: { type: string }
        meterId: { type: string }
        capacityKw: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        substation: { type: string, description: "ObjectId reference to this substation" }
    Reading:
      type: object
      properties:
        _id: { type: string }
        installation: { type: string }
        timestamp: { type: string, format: date-time }
        powerKw: { type: number }
        energyKwh: { type: number }
        voltage: { type: number }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            detail: { nullable: true }